requirements.txt
matplotlib=2.0.2=np113py36_0
numpy=1.13.3=py36ha266831_3
pandas=0.22.0=py36hf484d3e_0
pillow=4.1.1=py36_0
python=3.6.6=h6e4f718_2
pytorch=1.0.1=py3.6_cuda10.0.130_cudnn7.4.2_2
scikit-learn=0.20.3=py36hd81dba3_0
scipy=1.2.1=py36h7c811a0_0
torchvision=0.2.2=py_3
tqdm=4.31.1=pypi_0

